+++ ChangeLog.md
@@ -28,6 +28,9 @@ repository history](https://github.com/ddclient/ddclient/commits/main).
     [#907](https://github.com/ddclient/ddclient/pull/907)
   * Added `spaceship` protocol for Spaceship (https://www.spaceship.com/).
     [#896](https://github.com/ddclient/ddclient/pull/896)
+  * Added `asuscomm` protocol for ASUS DDNS (https://www.asuscomm.com/), with
+    support for both `asuscomm.com` and `asuscomm.cn` (mainland China).
+    [#909](https://github.com/ddclient/ddclient/pull/909)
 
 ## 2026-05-16 v4.0.1-rc.1